refactor/SafeAuth : Move the DisposableAction file into Models folder
